requirements.txt
fastapi
uvicorn
python-dotenv
openai
streamlit
pydantic
python-multipart
PyYAML
requests
beautifulsoup4
